Move GCSettings to shared (#22483)
Mainly just type definitions and error handling that gets shared.
