+++ src/System.Private.CoreLib/src/System/GC.cs
@@ -12,18 +12,10 @@
 **
 **
 ===========================================================*/
-//This class only static members and doesn't require the serializable keyword.
 
-using System;
-using System.Reflection;
-using System.Security;
-using System.Threading;
-using System.Runtime;
 using System.Runtime.CompilerServices;
-using System.Runtime.ConstrainedExecution;
 using System.Globalization;
 using System.Runtime.InteropServices;
-using System.Runtime.Versioning;
 using System.Diagnostics;
 
 namespace System
@@ -61,12 +53,6 @@ namespace System
     public static class GC
     {
         [MethodImplAttribute(MethodImplOptions.InternalCall)]
-        internal static extern int GetGCLatencyMode();
-
-        [MethodImplAttribute(MethodImplOptions.InternalCall)]
-        internal static extern int SetGCLatencyMode(int newLatencyMode);
-
-        [MethodImplAttribute(MethodImplOptions.InternalCall)]
         internal static extern void GetMemoryInfo(out uint highMemLoadThreshold,
                                                   out ulong totalPhysicalMem,
                                                   out uint lastRecordedMemLoad,
@@ -79,13 +65,7 @@ namespace System
 
         [DllImport(JitHelpers.QCall, CharSet = CharSet.Unicode)]
         internal static extern int _EndNoGCRegion();
-
-        [MethodImplAttribute(MethodImplOptions.InternalCall)]
-        internal static extern int GetLOHCompactionMode();
-
-        [MethodImplAttribute(MethodImplOptions.InternalCall)]
-        internal static extern void SetLOHCompactionMode(int newLOHCompactionMode);
-
+        
         [MethodImplAttribute(MethodImplOptions.InternalCall)]
         private static extern int GetGenerationWR(IntPtr handle);
 
@@ -100,9 +80,6 @@ namespace System
 
         [MethodImplAttribute(MethodImplOptions.InternalCall)]
         private static extern int _CollectionCount(int generation, int getSpecialGCCount);
-
-        [MethodImplAttribute(MethodImplOptions.InternalCall)]
-        internal static extern bool IsServerGC();
 
         [MethodImplAttribute(MethodImplOptions.InternalCall)]
         internal static extern ulong GetSegmentSize();
