+++ src/vm/ecalllist.h
@@ -776,13 +776,8 @@ FCFuncStart(gGCInterfaceFuncs)
     FCFuncElement("_WaitForFullGCComplete", GCInterface::WaitForFullGCComplete)
     FCFuncElement("_CollectionCount", GCInterface::CollectionCount)
     FCFuncElement("GetMemoryInfo", GCInterface::GetMemoryInfo)
-    FCFuncElement("GetGCLatencyMode", GCInterface::GetGcLatencyMode)
-    FCFuncElement("SetGCLatencyMode", GCInterface::SetGcLatencyMode)
-    FCFuncElement("GetLOHCompactionMode", GCInterface::GetLOHCompactionMode)
-    FCFuncElement("SetLOHCompactionMode", GCInterface::SetLOHCompactionMode)
     QCFuncElement("_StartNoGCRegion", GCInterface::StartNoGCRegion)
     QCFuncElement("_EndNoGCRegion", GCInterface::EndNoGCRegion)
-    FCFuncElement("IsServerGC", SystemNative::IsServerGC)
     FCFuncElement("GetSegmentSize", GCInterface::GetSegmentSize)
     QCFuncElement("_AddMemoryPressure", GCInterface::_AddMemoryPressure)
     QCFuncElement("_RemoveMemoryPressure", GCInterface::_RemoveMemoryPressure)
@@ -796,6 +791,14 @@ FCFuncStart(gGCInterfaceFuncs)
     FCFuncElement("_ReRegisterForFinalize", GCInterface::ReRegisterForFinalize)
 
     FCFuncElement("_GetAllocatedBytesForCurrentThread", GCInterface::GetAllocatedBytesForCurrentThread)
+FCFuncEnd()
+
+FCFuncStart(gGCSettingsFuncs)
+    FCFuncElement("get_IsServerGC", SystemNative::IsServerGC)
+    FCFuncElement("GetGCLatencyMode", GCInterface::GetGcLatencyMode)
+    FCFuncElement("GetLOHCompactionMode", GCInterface::GetLOHCompactionMode)
+    FCFuncElement("SetGCLatencyMode", GCInterface::SetGcLatencyMode)
+    FCFuncElement("SetLOHCompactionMode", GCInterface::SetLOHCompactionMode)
 FCFuncEnd()
 
 FCFuncStart(gInteropMarshalFuncs)
@@ -1224,6 +1227,7 @@ FCClassElement("FileLoadException", "System.IO", gFileLoadExceptionFuncs)
 FCClassElement("FormatterServices", "System.Runtime.Serialization", gSerializationFuncs)
 FCClassElement("GC", "System", gGCInterfaceFuncs)
 FCClassElement("GCHandle", "System.Runtime.InteropServices", gGCHandleFuncs)
+FCClassElement("GCSettings", "System.Runtime", gGCSettingsFuncs)
 #ifdef FEATURE_COMINTEROP
 FCClassElement("IEnumerable", "System.Collections", gStdMngIEnumerableFuncs)
 FCClassElement("IEnumerator", "System.Collections", gStdMngIEnumeratorFuncs)
